Inspection view completed
A few fixes to the inspection methods were taken care of. They weren't
checking for '.models' at the end of the app string causing them to
search for models at the module root rather than the module's models
submodule. Also, returning a list of tuples as Python's inspect module
would do is not optimal; we just need a list of names not the class
included.