Added condense methods and tests; fixed docstrings
Just added the last couple of methods to the Condenser class. Also fixed
a couple of docstrings to be accurate. Removed a few trailing spaces.
Simplified a few list comprehensions. Finally, commented out a test I
wrote while I was working on the views; it kept throwing an error, will
revisit it.